Cognito Import From Annotation
