+++ src/components/ImportFromCognitoS3Dialog/get-sources.js
@@ -1,0 +1,6 @@
+const getSources = (annotations) => {
+  return annotations
+    .map((obj) => obj.source)
+    .filter((source, index, self) => self.indexOf(source) === index)
+}
+export default getSources
