+++ src/components/ImportFromCognitoS3Dialog/set-type-of-file-to-load-and-disable.js
@@ -1,0 +1,38 @@
+import isEmpty from "lodash/isEmpty"
+import checkInterfaceAndAssets from "./check-interface-and-sample-type"
+export default (configImport, dataset) => {
+  return {
+    ...configImport,
+    typeOfFileToLoad:
+      !isEmpty(configImport) &&
+      !isEmpty(configImport.typeOfFileToLoad) &&
+      checkInterfaceAndAssets([configImport.typeOfFileToLoad, "Empty"], dataset)
+        ? configImport.typeOfFileToLoad
+        : checkInterfaceAndAssets(["Image", "Empty"], dataset)
+        ? "Image"
+        : checkInterfaceAndAssets(["Video", "Empty"], dataset)
+        ? "Video"
+        : checkInterfaceAndAssets(["Audio", "Empty"], dataset)
+        ? "Audio"
+        : checkInterfaceAndAssets(["PDF", "Empty"], dataset)
+        ? "PDF"
+        : checkInterfaceAndAssets(["Texte", "Empty"], dataset)
+        ? "Texte"
+        : "None",
+    typeOfFileToDisable: {
+      Image: checkInterfaceAndAssets(["Image", "Empty"], dataset)
+        ? false
+        : true,
+      Video: checkInterfaceAndAssets(["Video", "Empty"], dataset)
+        ? false
+        : true,
+      Audio: checkInterfaceAndAssets(["Audio", "Empty"], dataset)
+        ? false
+        : true,
+      PDF: checkInterfaceAndAssets(["PDF", "Empty"], dataset) ? false : true,
+      Texte: checkInterfaceAndAssets(["Texte", "Empty"], dataset)
+        ? false
+        : true,
+    },
+  }
+}
